move utils back to textelement
