+++ packages/excalidraw/element/textElement.ts
@@ -36,7 +36,6 @@ import {
   updateOriginalContainerCache,
 } from "./textWysiwyg";
 import { ExtractSetType } from "../utility-types";
-import { getBoundTextElementId } from "../../utils/textElement";
 
 export const normalizeText = (text: string) => {
   return (
@@ -942,4 +941,42 @@ export const getDefaultLineHeight = (fontFamily: FontFamilyValues) => {
     return DEFAULT_LINE_HEIGHT[fontFamily];
   }
   return DEFAULT_LINE_HEIGHT[DEFAULT_FONT_FAMILY];
+};
+
+export const getContainerElement = (
+  element:
+    | (ExcalidrawElement & {
+        containerId: ExcalidrawElement["id"] | null;
+      })
+    | null,
+  sceneElements: readonly NonDeletedExcalidrawElement[],
+) => {
+  if (!element) {
+    return null;
+  }
+  if (element.containerId) {
+    return sceneElements.find((ele) => ele.id === element.containerId);
+  }
+  return null;
+};
+
+export const getBoundTextElementId = (container: ExcalidrawElement | null) => {
+  return container?.boundElements?.length
+    ? container?.boundElements?.filter((ele) => ele.type === "text")[0]?.id ||
+        null
+    : null;
+};
+
+export const getBoundTextElement = (
+  element: ExcalidrawElement | null,
+  sceneElements: readonly NonDeletedExcalidrawElement[],
+) => {
+  if (!element) {
+    return null;
+  }
+  const boundTextElementId = getBoundTextElementId(element);
+  if (boundTextElementId) {
+    return sceneElements.find((ele) => ele.id === boundTextElementId);
+  }
+  return null;
 };
