+++ packages/utils/index.js
@@ -1,2 +1,1 @@
 export * from "./export";
-export * from "./textElement";
